Some code cleanup
git-svn-id: https://svn.apache.org/repos/asf/cxf/trunk@1393570 13f79535-47bb-0310-9956-ffa450edef68

+++ integration/jca/src/main/java/org/apache/cxf/jca/cxf/JCABusFactory.java
@@ -258,7 +258,7 @@ public class JCABusFactory {
                         LOG.info("ejbServicePropertiesFile modified, initializing/updating servants");
                         initializeServantsFromProperties(loadProperties(propsFile.toURI().toURL()));
                     }
-                    Thread.sleep(pollIntervalSeconds * 1000);
+                    Thread.sleep((long)pollIntervalSeconds * 1000L);
                 } catch (Exception e) {
                     LOG.info("MonitorThread: failed to initialiseServantsFromProperties "
                               + "with properties absolute path=" + propsFile.getAbsolutePath());
